Extracted search box to dedicated component
